+++ R/utilities.R
@@ -308,8 +308,13 @@ p
   font <- .parse_font(font.legend)
 
   if(!is.null(legend)) p <- p + theme(legend.position = legend)
-   p <- p +
-     labs(color = legend.title, fill = legend.title, linetype = legend.title, shape = legend.title)
+
+  if(!.is_empty(legend.title)){
+
+    if(.is_list(legend.title)) p <- p + do.call(ggplot2::labs, legend.title)
+    else p <- p +
+       labs(color = legend.title, fill = legend.title, linetype = legend.title, shape = legend.title)
+  }
 
    if(!is.null(font)){
      p <- p + theme(